backend-fastapi/requirements.txt
bcrypt
uvicorn[standard]
fastapi
sqlalchemy
pydantic
typing




# Generowanie muzyki - MIDI
mido
music21
pretty_midi

# Audio processing
numpy
scipy
pydub

# Sample fetching
requests
matplotlib
python-dotenv
openai
anthropic
google-generativeai
python-jose
